# trying again: Big performance gains for Go NewTensor and Value