This is re-establishing PR https://github.com/tensorflow/tensorflow/pull/39117 after some fat-fingering. Should be exactly the same as that PR

After some helpful comments in #36578 (comment) I'm trying again with this performance improvement.

Please please please if it gets rolled back again because of "internal test" failures, please get some kind of debug information. Stack traces or something. Any kind of clue.

```
name                                  old time/op    new time/op    delta
Tensor/New/[150528]int32-16             1.78ms ± 4%    0.13ms ±10%   -92.63%  (p=0.000 n=8+7)
Tensor/New/[100][100][100]int32-16      13.1ms ± 1%     0.9ms ±53%   -92.81%  (p=0.000 n=8+8)
Tensor/New/[]float32-16                 3.72ms ± 1%    0.97ms ±30%   -74.04%  (p=0.000 n=8+8)
Tensor/New/[][]float32-16               4.83ms ± 2%    1.32ms ± 8%   -72.69%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16             4.81ms ± 1%    1.32ms ± 4%   -72.51%  (p=0.001 n=8+6)
Tensor/New/[]string-16                   466ms ± 1%      34ms ± 4%   -92.60%  (p=0.001 n=7+7)
Tensor/New/[][]string-16                 460ms ± 1%      35ms ± 1%   -92.45%  (p=0.000 n=8+8)
Tensor/New/[][][]string-16               462ms ± 2%      36ms ± 5%   -92.14%  (p=0.000 n=8+8)
Tensor/Value/[150528]int32-16            647µs ± 3%      82µs ± 1%   -87.28%  (p=0.000 n=8+8)
Tensor/Value/[100][100][100]int32-16    6.43ms ± 1%    0.99ms ± 3%   -84.63%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16               5.57ms ± 3%    1.04ms ± 7%   -81.26%  (p=0.000 n=8+8)
Tensor/Value/[][]float32-16             6.84ms ± 1%    1.51ms ± 1%   -77.96%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16           6.87ms ± 1%    1.52ms ± 3%   -77.80%  (p=0.001 n=7+7)
Tensor/Value/[]string-16                 268ms ± 3%      20ms ± 2%   -92.45%  (p=0.000 n=8+8)
Tensor/Value/[][]string-16               269ms ± 2%      20ms ± 1%   -92.46%  (p=0.000 n=8+7)
Tensor/Value/[][][]string-16             271ms ± 2%      20ms ± 1%   -92.55%  (p=0.000 n=8+8)

name                                  old alloc/op   new alloc/op   delta
Tensor/New/[150528]int32-16              606kB ± 0%       0kB ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/New/[100][100][100]int32-16      4.16MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=7+8)
Tensor/New/[]float32-16                 4.01MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.002 n=7+8)
Tensor/New/[][]float32-16               4.48MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16             4.48MB ± 0%    0.00MB ± 0%  -100.00%  (p=0.002 n=7+8)
Tensor/New/[]string-16                  48.0MB ± 0%     0.0MB ± 0%  -100.00%  (p=0.000 n=7+8)
Tensor/New/[][]string-16                48.3MB ± 0%     0.0MB ± 0%  -100.00%  (p=0.000 n=7+8)
Tensor/New/[][][]string-16              48.3MB ± 0%     0.0MB ± 0%  -100.00%  (p=0.000 n=7+8)
Tensor/Value/[150528]int32-16           1.21MB ± 0%    0.61MB ± 0%   -50.00%  (p=0.000 n=8+8)
Tensor/Value/[100][100][100]int32-16    9.23MB ± 0%    4.25MB ± 0%   -53.93%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16               8.01MB ± 0%    4.01MB ± 0%   -50.00%  (p=0.000 n=8+7)
Tensor/Value/[][]float32-16             9.21MB ± 0%    4.25MB ± 0%   -53.82%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16           9.23MB ± 0%    4.25MB ± 0%   -53.93%  (p=0.000 n=8+8)
Tensor/Value/[]string-16                56.0MB ± 0%    23.0MB ± 0%   -58.91%  (p=0.000 n=8+7)
Tensor/Value/[][]string-16              58.5MB ± 0%    23.3MB ± 0%   -60.23%  (p=0.000 n=8+8)
Tensor/Value/[][][]string-16            58.5MB ± 0%    23.3MB ± 0%   -60.25%  (p=0.001 n=7+7)

name                                  old allocs/op  new allocs/op  delta
Tensor/New/[150528]int32-16               4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=8+8)
Tensor/New/[100][100][100]int32-16       10.0k ± 0%      0.0k ± 0%   -99.96%  (p=0.000 n=8+8)
Tensor/New/[]float32-16                   4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=8+8)
Tensor/New/[][]float32-16                20.0k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/New/[][][]float32-16              20.0k ± 0%      0.0k ± 0%   -99.98%  (p=0.000 n=8+8)
Tensor/New/[]string-16                   4.00M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/New/[][]string-16                 4.01M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/New/[][][]string-16               4.01M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/Value/[150528]int32-16             7.00 ± 0%      2.00 ± 0%   -71.43%  (p=0.000 n=8+8)
Tensor/Value/[100][100][100]int32-16     40.2k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/Value/[]float32-16                 7.00 ± 0%      2.00 ± 0%   -71.43%  (p=0.000 n=8+8)
Tensor/Value/[][]float32-16              40.0k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/Value/[][][]float32-16            40.2k ± 0%      0.0k ± 0%   -99.99%  (p=0.000 n=8+8)
Tensor/Value/[]string-16                 5.00M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/Value/[][]string-16               5.02M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)
Tensor/Value/[][][]string-16             5.02M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=8+8)```